server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Shapely==2.0.2
requests==2.31.0
gunicorn==21.2.0
